Add files via upload
Took a while but I prevailed. I was able to recreate most of this code in 30 minutes because of practice. however, I did struggle with the deletion of the node because I forgot to reset the temp to the next pointer. I used a pointer to pointer reference to finished off head Node  Originally, I put the node in a class but decided that it was better to just use a structer because Classes  automatically private in C++ while structures are public.