Update main.cpp
I got ride of the reference pointer. I added correct the delete algorithm so that the List can delete all the numbers one by one. I modified the code so that it wont crash if you search for the same number twice by accident. The only thing missing is the return my value function then it will be complete.